pyproject.toml
[tool.poetry]
name = "best-picture-nominees-davidm3"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
imdb-cli-tool = "^1.0.1"
tabulate = "^0.8.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
